+++ amazon/ec2credentials/fake/src/main/kotlin/org/http4k/connect/amazon/ec2credentials/FakeEc2CredentialService.kt
@@ -4,7 +4,6 @@ import org.http4k.base64Encode
 import org.http4k.chaos.ChaoticHttpHandler
 import org.http4k.chaos.start
 import org.http4k.connect.amazon.core.model.AccessKeyId
-import org.http4k.connect.amazon.core.model.Ec2Credentials
 import org.http4k.connect.amazon.core.model.Ec2ProfileName
 import org.http4k.connect.amazon.core.model.Expiration
 import org.http4k.connect.amazon.core.model.SecretAccessKey