+++ CHANGELOG.md
@@ -5,6 +5,8 @@ changes with their rationale when appropriate. Given version `A.B.C.D`, breaking
 
 ### v3.26.3.0 (uncut)
 - **http4k-connect-*** - Upgrade dependencies
+- **http4k-connect-amazon-instancemetadata** - New module!  Query metadata and credentials from the current Amazon EC2 environment.  H/T @oharaandrew314
+- **http4k-connect-amazon-ec2credentials** - Deprecated.  Use the **http4k-connect-amazon-instancemetadata** module
 
 ### v3.26.3.0
 - **http4k-connect-cognito** - We now generate action code using Kapt, as per the other adapters.